[level 2] Title: k진수에서 소수 개수 구하기, Time: 0.09 ms, Memory: 82.3 MB -BaekjoonHub

+++ 프로그래머스/2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -13,7 +13,7 @@ class Solution {
 
     }
 
-    public long solution(long N, long K) {
+    public int solution(int N, int K) {
 
 
         if (N == 0) {
